+++ js/check-string-length.js
@@ -1,0 +1,2 @@
+const checkStringLength = (string = '', maxSymbols = 1) => string.length <= maxSymbols;
+export {checkStringLength};
