+++ js/make-counter.js
@@ -1,0 +1,12 @@
+const makeCounter = () => {
+  let currentCount = 1;
+
+  return function() {
+    return currentCount++;
+  };
+};
+
+const Id = makeCounter();
+const Photoid = makeCounter();
+
+export {Id, Photoid};
